@@ -1,115 +1,70 @@
 {
   "name": "magic-box",
-  "version": "v1.22.10",
+  "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@date-io/moment": "^2.10.6",
-    "@formatjs/intl-relativetimeformat": "^4.5.16",
-    "@fortawesome/fontawesome-pro": "^5.14.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.30",
-    "@fortawesome/free-brands-svg-icons": "^5.14.0",
-    "@fortawesome/free-regular-svg-icons": "^5.14.0",
-    "@fortawesome/free-solid-svg-icons": "^5.14.0",
-    "@fortawesome/pro-duotone-svg-icons": "^5.14.0",
-    "@fortawesome/pro-light-svg-icons": "^5.14.0",
-    "@fortawesome/pro-regular-svg-icons": "^5.14.0",
-    "@fortawesome/pro-solid-svg-icons": "^5.14.0",
-    "@fortawesome/react-fontawesome": "^0.1.11",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@material-ui/pickers": "^3.2.10",
-    "a2hs": "^1.1.4",
-    "axios": "^0.19.2",
-    "chart.js": "^2.9.3",
-    "downshift": "^5.4.7",
-    "firebase": "^7.21.0",
-    "firebaseui": "^4.6.1",
-    "fireform": "^1.0.0",
-    "firekit": "^1.0.34",
-    "firekit-provider": "^1.0.2",
-    "github-markdown-css": "^4.0.0",
-    "history": "^4.10.1",
-    "immutable": "^4.0.0-rc.12",
-    "intl": "^1.2.5",
-    "intl-locales-supported": "^1.8.12",
-    "intl-pluralrules": "^1.2.2",
-    "jwt-decode": "^3.1.1",
-    "lazysizes": "^5.2.0",
-    "match-sorter": "^4.2.1",
-    "material-ui-filter": "^3.1.3",
-    "material-ui-image": "^3.2.4",
-    "material-ui-pickers": "^2.2.4",
-    "mathjs": "^7.6.0",
-    "moment": "^2.29.0",
-    "muishift": "^1.1.4",
-    "prop-types": "^15.7.2",
-    "qrcode.react": "^1.0.0",
-    "react": "^16.13.1",
-    "react-chartjs-2": "^2.10.0",
-    "react-countup": "^4.3.3",
-    "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^16.13.1",
-    "react-dropzone": "^11.2.0",
-    "react-firebaseui": "^4.1.0",
-    "react-helmet": "^5.2.1",
-    "react-image-cropper": "^1.3.0",
-    "react-intl": "^5.7.1",
-    "react-list": "^0.8.15",
-    "react-loadable": "^5.5.0",
-    "react-markdown": "^4.3.1",
-    "react-mic": "^12.4.6",
-    "react-multi-carousel": "^2.5.5",
-    "react-on-screen": "^2.1.1",
-    "react-player": "^2.6.2",
-    "react-redux": "^7.2.1",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.1",
-    "react-syntax-highlighter": "^15.4.3",
-    "react-toastify": "^6.0.8",
-    "react-virtualized": "^9.22.2",
-    "redux": "^4.0.5",
-    "redux-form": "^8.3.6",
-    "redux-form-material-ui": "^5.0.0-beta.3",
-    "redux-logger": "^3.0.6",
-    "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.3.0",
-    "rmw-shell": "^7.2.1",
-    "testcafe": "^1.8.4"
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/material": "^5.11.5",
+    "@mui/styles": "^5.11.2",
+    "@types/node": "^17.0.45",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "jwt-decode": "^3.1.2",
+    "mathjs": "^11.5.0",
+    "qrcode.react": "^3.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "react-syntax-highlighter": "^15.5.0",
+    "typescript": "^4.9.4",
+    "web-vitals": "^2.1.4",
+    "workbox-background-sync": "^6.5.4",
+    "workbox-broadcast-update": "^6.5.4",
+    "workbox-cacheable-response": "^6.5.4",
+    "workbox-core": "^6.5.4",
+    "workbox-expiration": "^6.5.4",
+    "workbox-google-analytics": "^6.5.4",
+    "workbox-navigation-preload": "^6.5.4",
+    "workbox-precaching": "^6.5.4",
+    "workbox-range-requests": "^6.5.4",
+    "workbox-routing": "^6.5.4",
+    "workbox-strategies": "^6.5.4",
+    "workbox-streams": "^6.5.4"
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@typescript-eslint/eslint-plugin": "^5.13.0",
+    "@typescript-eslint/parser": "^5.0.0",
+    "eslint": "^8.18.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^17.0.0"
   },
   "scripts": {
-    "check": "npm-check -s -u",
-    "analyze": "source-map-explorer build/static/js/main.*",
-    "start": "HOST=0.0.0.0 PORT=8085 react-scripts start",
-    "standard": "standard src/**/*.js ",
+    "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-  "author": "Tarik Huber & Xiao Xiao",
-  "homepage": "https://www.react-most-wanted.com/",
-  "license": "MIT",
-  "repository": "https://github.com/TarikHuber/react-most-wanted",
-  "keywords": [
-    "react-most-wanted",
-    "react",
-    "redux",
-    "firebase"
-  ],
-  "prettier": {
-    "trailingComma": "es5",
-    "tabWidth": 2,
-    "semi": false,
-    "singleQuote": true
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
-  "devDependencies": {
-    "selenium-webdriver": "^4.0.0-alpha.5"
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
