@@ -1,13 +1,14 @@
 /* eslint-disable */
-import React, { } from 'react'
-import { injectIntl } from 'react-intl'
-import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles'
-import { Activity } from 'rmw-shell'
-import { Container, Grid, TextField, AppBar, Toolbar, Typography } from '@material-ui/core'
-import MagicBox from './magicBox'
+import React, { useEffect, useState } from 'react'
+import { AppBar, Container, Grid, TextField, Toolbar, Typography } from '@mui/material'
+import { createTheme } from '@mui/material/styles';
+import { ThemeProvider } from '@mui/styles'
+import MagicBox from './MagicBox'
 
 
-const theme = createMuiTheme({
+const MagicBoxTitle = 'Magic Box'
+
+const theme = createTheme({
   overrides: {
     MuiInput: {
       textarea: {
@@ -17,28 +18,27 @@ const theme = createMuiTheme({
   },
 })
 
-const MagicBoxPage = ({intl}) => {
-  const MagicBoxTitle = 'Magic Box'
-  const [userInput, setUserInput] = React.useState('')
-  const [magicIn, setMagicIn] = React.useState('')
+const MagicBoxPage = () => {
+  const [userInput, setUserInput] = useState('')
+  const [magicIn, setMagicIn] = useState('')
 
-  React.useEffect(() => {
+  useEffect(() => {
     const timeOutID = setTimeout(() => setMagicIn(userInput), 500);
     return () => { clearTimeout(timeOutID) }
   }, [userInput]);
 
   // style="height: 50px; padding-right: 10px"
   return (
-    <MuiThemeProvider theme={theme}>
+    <ThemeProvider theme={theme}>
 
-      <AppBar position="static" style={{
+      <AppBar position="static" sx={{
           color: theme.palette.text.secondary,
           backgroundColor: 'white',
           borderBottom: '1px solid #cdc9c3',
           height: "65px",
        }}>
         <Toolbar>
-          <img src="/images/cat-50.png" alt="icon" style={{
+          <img src="/images/logo-50.png" alt="icon" style={{
             height: "50px",
             paddingRight: "10px",
           }} />
@@ -46,9 +46,8 @@ const MagicBoxPage = ({intl}) => {
         </Toolbar>
       </AppBar>
 
-      <Container maxWidth="xl">
-        <Grid container spacing={2}
-            style={{ marginTop: "30px", padding: "30px" }}>
+      <Container maxWidth="95vw">
+        <Grid container spacing={2} sx={{ marginTop: "30px", padding: "30px" }}>
 
           {/* Magic Box Input */}
           <Grid item xs={12} sm={12} md={6} >
@@ -59,19 +58,22 @@ const MagicBoxPage = ({intl}) => {
               fullWidth={true}
               rows={7}
               onChange={(e) => { setUserInput(e.target.value) }}
-              style={{ fontSize: "1.25rem" }} // this style attr not work
-              variant={"outlined"}
+              sx={{ fontSize: "1.25rem" }} // this style attr not work
+              variant='outlined'
             />
           </Grid>
 
           {/* Magic Box Output */}
-          <MagicBox in={magicIn} />
+          <Grid container item xs={12} sm={12} md={6}
+            sx={{ background: "#f5f5f5", overflow: "scroll", maxHeight: "calc(100vh - 100px)" }} >
+            <MagicBox in={magicIn} />
+          </Grid>
 
         </Grid>
       </Container>
 
 
-      <AppBar position="fixed" style={{
+      <AppBar position="fixed" sx={{
           top: 'auto',
           bottom: 0,
           color: theme.palette.text.secondary,
@@ -83,8 +85,8 @@ const MagicBoxPage = ({intl}) => {
         {`© ${new Date().getFullYear()} Copyright: All Rights Not Reserved`}
       </AppBar>
 
-    </MuiThemeProvider>
+    </ThemeProvider>
   )
 }
 
-export default injectIntl(MagicBoxPage)
+export default MagicBoxPage
